folder restructure for draggables
