+++ packages/easy-email-extensions/src/KiaLayout/draggables/content/AccordianDraggable.tsx
@@ -1,6 +1,6 @@
 import { BasicType } from 'easy-email-core';
 import React from 'react';
-import { DraggableWrapper } from './DraggableWrapper';
+import { DraggableWrapper } from '../DraggableWrapper';
 import { getImg } from '@extensions/ShortcutToolbar/utils/getImg';
 
 interface IAccordianDraggableProps {
