+++ packages/easy-email-extensions/src/KiaLayout/draggables/content/CarouselDraggable.tsx
@@ -1,8 +1,8 @@
 import { BasicType } from 'easy-email-core';
 import React from 'react';
-import { DraggableWrapper } from './DraggableWrapper';
+import { DraggableWrapper } from '../DraggableWrapper';
 import { getImg } from '@extensions/ShortcutToolbar/utils/getImg';
-import Placeholder from '../assets/Placeholder.svg';
+import Placeholder from '../../assets/Placeholder.svg';
 
 interface ICarouselDraggableProps {
   children: React.ReactNode;
