+++ packages/easy-email-extensions/src/KiaLayout/draggables/content/DividerDraggable.tsx
@@ -1,6 +1,6 @@
 import { BasicType, IDivider, RecursivePartial } from 'easy-email-core';
 import React from 'react';
-import { DraggableWrapper } from './DraggableWrapper';
+import { DraggableWrapper } from '../DraggableWrapper';
 
 interface IDividerDraggableProps {
   children: React.ReactNode;
